Rename workflow package to workflows
