+++ internal/workflows/preprocessing.go
@@ -1,4 +1,4 @@
-package workflow
+package workflows
 
 import (
 	"fmt"
