+++ internal/workflows/preprocessing_test.go
@@ -1,4 +1,4 @@
-package workflow_test
+package workflows_test
 
 import (
 	"fmt"
@@ -14,7 +14,7 @@ import (
 	temporalsdk_worker "go.temporal.io/sdk/worker"
 
 	"github.com/artefactual-sdps/custom-enduro-workflows/internal/config"
-	"github.com/artefactual-sdps/custom-enduro-workflows/internal/workflow"
+	"github.com/artefactual-sdps/custom-enduro-workflows/internal/workflows"
 )
 
 const sharedPath = "/shared/path/"
@@ -24,7 +24,7 @@ type PreprocessingTestSuite struct {
 	temporalsdk_testsuite.WorkflowTestSuite
 
 	env      *temporalsdk_testsuite.TestWorkflowEnvironment
-	workflow *workflow.PreprocessingWorkflow
+	workflow *workflows.PreprocessingWorkflow
 }
 
 func (s *PreprocessingTestSuite) SetupTest(cfg config.Configuration) {
@@ -38,7 +38,7 @@ func (s *PreprocessingTestSuite) SetupTest(cfg config.Configuration) {
 	)
 
 	cfg.Preprocessing.SharedPath = sharedPath
-	s.workflow = workflow.NewPreprocessingWorkflow(cfg.Preprocessing)
+	s.workflow = workflows.NewPreprocessingWorkflow(cfg.Preprocessing)
 }
 
 func (s *PreprocessingTestSuite) AfterTest(suiteName, testName string) {
